Include overview

+++ database/entity/ping.go
@@ -3,6 +3,6 @@ package entity
 type PingEntity struct {
 	ServiceName  string `json:"serviceName"`
 	ResponseTime int    `json:"responseTime"`
-	CreatedAt    int    `json:"createdAt"`
-	UpdatedAt    int    `json:"updatedAt"`
+	CreatedAt    int64  `json:"createdAt"`
+	UpdatedAt    int64  `json:"updatedAt"`
 }
